refactor: ensure image creation

+++ registration/models.py
@@ -1,13 +1,14 @@
 from django.contrib.auth.models import User
 from django.db import models
-from django.db.models.signals import pre_save
+from django.db.models.signals import post_save
 from django.dispatch import receiver
 
 
 def custom_uplad_to(instance, filename):
-    old_instance = Profile.objects.get(pk=instance.pk)
-    old_instance.avatar.delete()
-    return 'profiles/' + filename
+        if Profile.objects:
+            old_instance = Profile.objects.get(pk=instance.pk)
+            old_instance.avatar.delete()
+        return 'profiles/' + filename
 
 
 class Profile(models.Model):
@@ -32,7 +33,7 @@ class Profile(models.Model):
     )
 
 
-@receiver(pre_save, sender=User)
+@receiver(post_save, sender=User)
 def ensure_profile_creation(sender, instance, **kwargs):
     if kwargs.get('created', False):
         Profile.objects.get_or_create(user=instance)
